requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
lxml==4.9.3
aiosqlite==0.19.0
tiktoken==0.5.2
unidecode==1.3.7
pyyaml==6.0.1
pgvector==0.3.0
striprtf==0.0.28

